test(outputs.graylog): skip tcp/udp tests on windows & darwin
This test is constantly timing out on darwin and at times windows
systems in circleci.
